middleware authorization

+++ routes.php
@@ -3,16 +3,20 @@
 $router->get('/', 'HomeController@index');
 
 $router->get('/listings', 'ListingController@index');
-$router->post('/listings', 'ListingController@store');
+$router->post('/listings', 'ListingController@store', ['auth']);
 
-$router->get('/listings/create', 'ListingController@create');
+$router->get('/listings/create', 'ListingController@create', ['auth']);
 
-$router->get('/listings/edit/{id}', 'ListingController@edit');
+$router->get('/listings/edit/{id}', 'ListingController@edit', ['auth']);
+
 $router->get('/listings/{id}', 'ListingController@show');
-$router->put('/listings/{id}', 'ListingController@update');
-$router->delete('/listings/{id}', 'ListingController@destroy');
+$router->put('/listings/{id}', 'ListingController@update', ['auth']);
+$router->delete('/listings/{id}', 'ListingController@destroy', ['auth']);
 
-$router->get('/auth/register', 'UserController@create');
-$router->get('/auth/login', 'UserController@login');
+$router->get('/auth/register', 'UserController@create', ['guest']);
+$router->post('/auth/register', 'Usercontroller@store', ['guest']);
 
-$router->post('/auth/register', 'Usercontroller@store');
+$router->get('/auth/login', 'UserController@login', ['guest']);
+$router->post('/auth/login', 'Usercontroller@authenticate');
+
+$router->post('/auth/logout', 'Usercontroller@logout', ['auth']);
